SEI: TicTacToe Project
This is the very beginning of my first project at GA Boston where I will
be making the classic game Tic Tac Toe from scratch where the user will
play against themselves. The game will consist of HTML, CSS and
JavaScript files.

The user should be able to sign up with an email, password and
confirming the password to be able to play the game. The user
should be able to sign in to play with the email and password they
created and sign out of the game. They should be able to change their
password as well and most importantly the player should be able to
play the game smoothly.

The user should be able to switch turns between X and O on a game board
and the application should display which side won or if there was
a draw. An API will be used to create a new game, update the game by
storing new moves, and visually display game statistics like how many
games the user won.

This is the first commit done on 09/25/19.