SEI: Tic-Tac-Toe Project 09/28/19
I currently have the Authentication files set up and are working. When
the user is on main page they should see the sign in form and the sign
up form. When the user signs up they can then sign in to play. When the
user signs in they should see the game, along with the the change
password form and the sign out button.