SEI: Tic-Tac-Toe
I was able to get some style on the page but I still need to work on
determining a tie and displaying it to the user. I also need to show the
user the game stats like showing the number of times a player has won
and the number of games played.